[DEV] Fix/about (#27)
* style: Change scrollbar thumb bg color

* chore: Remove file in order change the route

* chore: Modify and Update import name

* chore(module): Separate data for it's uses and purposes

* chore: Update route into private route by adding underscore prefix at partials directory

* chore(route): Move chatbot route inside dashboard route

* chore(route): Move about directory inside landing partial's directory

* chore(ui): Modify style for layout grid breakpoint

* style(component): Modify bg color of hover state

* chore(component): Modify rating button bg color

* chore: Modify footer content

* feat(component): Create skeleton component for image grid layout
