+++ app/(route)/dashboard/chatbot/page.tsx
