+++ app/(route)/landing/page.tsx
@@ -1,7 +1,7 @@
 import { FloatingNav } from '@/components/ui/navigation/FloatingNav'
 import { NextPage } from 'next'
-import { landingLinks } from '@/data/index'
-import About from '@/app/(route)/about/page'
+import { landingLinks } from '@/data/landing-link'
+import About from '@/app/(route)/landing/_partials/about/page'
 import HeroBanner from '@/app/(route)/landing/_partials/hero-banner/page'
 import Recommendation from '@/app/(route)/landing/_partials/recomendation/page'
 import OurService from '@/app/(route)/landing/_partials/our-services/page'
