+++ app/globals.css
@@ -18,7 +18,7 @@
     background: transparent;
   }
   ::-webkit-scrollbar-thumb {
-    background: #000;
+    background: #155e75;
   }
 
   scroll-behavior: smooth;
