+++ components/ui/card/CardHoverEffect.tsx
@@ -36,7 +36,7 @@ export const HoverEffect = ({
           <AnimatePresence>
             {hoveredIndex === idx && (
               <motion.span
-                className='absolute inset-0 block h-full w-full rounded-3xl bg-cyan-300/[.8] dark:bg-slate-800/[0.8]'
+                className='absolute inset-0 block h-full w-full rounded-3xl bg-cyan-600/[.8] dark:bg-slate-800/[0.8]'
                 layoutId='hoverBackground'
                 initial={{ opacity: 0 }}
                 animate={{
