+++ data/index.ts
@@ -1,0 +1,83 @@
+export const heroImageSliderImages = [
+  'https://ik.imagekit.io/rom/melancong/landscape/image-2-landscape.webp?updatedAt=1721609667633',
+  'https://ik.imagekit.io/rom/melancong/landscape/image-8-landscape.webp?updatedAt=1721609671401',
+  'https://ik.imagekit.io/rom/melancong/landscape/image-12-landscape.webp?updatedAt=1721609668098'
+]
+
+export const words = ['discover', 'explore', 'find', 'Enjoy']
+
+export const services = [
+  {
+    title: 'Destination Highlights',
+    description:
+      'Discover the most popular tourist destinations in Bali, featuring top attractions and must-see spots.',
+    link: 'https://stripe.com'
+  },
+  {
+    title: 'Travel Guides',
+    description:
+      'Comprehensive travel guides to help you navigate Bali’s rich culture, beautiful landscapes, and hidden gems.',
+    link: 'https://netflix.com'
+  },
+  {
+    title: 'Activity Recommendations',
+    description:
+      'Get personalized activity recommendations, from adventure sports to cultural experiences, tailored to your interests.',
+    link: 'https://google.com'
+  },
+  {
+    title: 'Dining Suggestions',
+    description:
+      'Explore the culinary delights of Bali with our curated list of the best restaurants, cafes, and local eateries.',
+    link: 'https://meta.com'
+  },
+  {
+    title: 'Travel Planning Tools',
+    description:
+      'Utilize our suite of planning tools, including itineraries, maps, and budget calculators, to make your Bali trip seamless and enjoyable.',
+    link: 'https://amazon.com'
+  },
+  {
+    title: 'Accommodation Tips',
+    description:
+      'Find the best places to stay, from luxury resorts to budget-friendly accommodations, all with detailed reviews and ratings.',
+    link: 'https://microsoft.com'
+  }
+]
+
+export const destinationLocations = [
+  'Tabanan',
+  'Badung',
+  'Gianyar',
+  'Denpasar',
+  'Karangasem',
+  'Buleleng',
+  'Klungkung',
+  'Bangli',
+  'Jembrana'
+]
+
+export const destinationTypes = [
+  'Historical Landmark',
+  'Beach',
+  'Temple',
+  'Wildlife',
+  'Museum',
+  'Garden',
+  'Lake',
+  'Waterfall',
+  'Mountain',
+  'Hot Spring',
+  'Rice Field',
+  'Culture',
+  'Hill',
+  'Countryside'
+]
+
+export const inputSearchPlaceholders = [
+  'Searching for destinations?',
+  'Try to search for tanah lot..',
+  'Not sure where to go?',
+  'Try to search for the beach',
+  'Not solved your issue? then try to ask for our chatbot'
+]
